Converted all double calculations to float Optimized matrix mutliplications

+++ DenseNeuralNetwork/Activation.h
@@ -19,8 +19,8 @@ class Activation {
 		static Activation* ALL_ACTIVATIONS[NUM_ACTIVATIONS];
 
 		bool condenseGradient = true;
-		virtual void operate(int batchSize, int size, double** activations, double** neurons) = 0;
-		virtual void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient) = 0;
+		virtual void operate(int batchSize, int size, float** activations, float** neurons) = 0;
+		virtual void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient) = 0;
 		virtual Activation* clone() = 0;
 		virtual bool isDiagonal() { return true; };
 		virtual void save(ofstream& file) {
@@ -33,66 +33,66 @@ class Activation {
 class None : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Sigmoid : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Relu : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Elu : public Activation {
 
 public:
-	double alpha;
-	Elu(double alpha);
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	float alpha;
+	Elu(float alpha);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 	void save(ofstream& file);
 };
 
 class Selu : public Activation {
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Tanh : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Swish : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 };
 
 class Softmax : public Activation {
 
 public:
-	void operate(int batchSize, int size, double** activations, double** neurons);
-	void differentiate(int batchSize, int size, double** activations, double** neurons, double*** activationGradient);
+	void operate(int batchSize, int size, float** activations, float** neurons);
+	void differentiate(int batchSize, int size, float** activations, float** neurons, float*** activationGradient);
 	Activation* clone();
 	bool isDiagonal();
 };
